use size_t for buffers in libcurl internal impl (#2376)
* use size_t for buffers in libcurl internal impl

+++ sdk/core/azure-core/src/http/curl/curl_connection_pool_private.hpp
@@ -119,7 +119,7 @@ namespace Azure { namespace Core { namespace Http { namespace _detail {
 
     // Makes possible to know the number of current connections in the connection pool for an
     // index
-    int64_t ConnectionsOnPool(std::string const& host) { return ConnectionPoolIndex[host].size(); };
+    size_t ConnectionsOnPool(std::string const& host) { return ConnectionPoolIndex[host].size(); };
 
     std::thread m_cleanThread;
   };
